+++ etc/Jenkinsfile
@@ -5,7 +5,9 @@ pipeline {
 		stage('Publish Data') {
 			steps {
 				checkout scm
-				publishReport pattern: "etc/*.json", displayType: "dual"
+				publishReport name: "JSON Report", displayType: "dual", provider: json(pattern: "etc/*.json")
+				publishReport name: "XML Report", displayType: "dual", provider: xml(pattern: "etc/*.xml")
+				publishReport name: "YAML Report", displayType: "dual", provider: yaml(pattern: "etc/*.yaml")
 			} 
 		}
 	}
