--- etc/Jenkinsfile.new
@@ -1,14 +1,0 @@
-pipeline {
-	agent any
-	
-	stages {
-		stage('Publish Data') {
-			steps {
-				checkout scm
-				publishReports name: "JSON Report", displayType: "dual", provider: json(pattern: "etc/*.json")
-				publishReports name: "XML Report", displayType: "dual", provider: xml(pattern: "etc/*.xml")
-				publishReports name: "YAML Report", displayType: "dual", provider: yaml(pattern: "etc/*.yaml")
-			} 
-		}
-	}
-}
