+++ src/main/java/io/jenkins/plugins/reporter/model/ItemTableModel.java
@@ -2,7 +2,7 @@ package io.jenkins.plugins.reporter.model;
 
 import io.jenkins.plugins.datatables.TableColumn;
 import io.jenkins.plugins.reporter.ItemViewModel;
-import io.jenkins.plugins.reporter.steps.Item;
+import io.jenkins.plugins.reporter.nextgen.Item;
 import org.apache.commons.text.CaseUtils;
 
 import java.util.ArrayList;
