+++ src/main/java/io/jenkins/plugins/reporter/model/Result.java
@@ -3,8 +3,8 @@ package io.jenkins.plugins.reporter.model;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.jenkins.plugins.reporter.steps.Item;
-import io.jenkins.plugins.reporter.steps.Palette;
+import io.jenkins.plugins.reporter.nextgen.Item;
+import io.jenkins.plugins.reporter.nextgen.Palette;
 
 import java.io.Serializable;
 import java.util.ArrayList;
