+++ src/main/java/io/jenkins/plugins/reporter/nextgen/Palette.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.reporter.steps;
+package io.jenkins.plugins.reporter.nextgen;
 
 public enum Palette {
     YELLOW("#FFF59D"),
