+++ src/main/java/io/jenkins/plugins/reporter/nextgen/provider/Xml.java
@@ -1,28 +1,32 @@
-package io.jenkins.plugins.reporter.steps.provider;
+package io.jenkins.plugins.reporter.nextgen.provider;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.xml.XmlFactory;
 import edu.umd.cs.findbugs.annotations.CheckForNull;
 import hudson.Extension;
-import io.jenkins.plugins.reporter.steps.ReportDto;
-import io.jenkins.plugins.reporter.steps.Provider;
-import io.jenkins.plugins.reporter.steps.ReportParser;
+import io.jenkins.plugins.reporter.nextgen.ReportDto;
+import io.jenkins.plugins.reporter.nextgen.Provider;
+import io.jenkins.plugins.reporter.nextgen.ReportParser;
 import org.apache.commons.lang3.StringUtils;
 import org.jenkinsci.Symbol;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 
 import java.io.File;
+import java.io.IOException;
 
-public class Csv extends Provider {
+public class Xml extends Provider {
     
     private static final long serialVersionUID = 9141170397250309265L;
 
-    private static final String ID = "csv";
-    private String pattern = StringUtils.EMPTY;
-
-    private String name = StringUtils.EMPTY;
+    private static final String ID = "xml";
     
+    private String pattern = StringUtils.EMPTY;
+    
+    private String name = StringUtils.EMPTY;
+
     @DataBoundConstructor
-    public Csv() {
+    public Xml() {
         super();
         // empty constructor required for stapler
     }
@@ -55,11 +59,11 @@ public class Csv extends Provider {
 
     @Override
     public ReportParser createParser() {
-        return new CsvParser();
+        return new XmlParser();
     }
 
     /** Descriptor for this provider. */
-    @Symbol("csv")
+    @Symbol("xml")
     @Extension
     public static class Descriptor extends ProviderDescriptor {
         /** Creates the descriptor instance. */
@@ -67,14 +71,14 @@ public class Csv extends Provider {
             super(ID);
         }
     }
+    
+    public static class XmlParser extends ReportParser {
 
-    public static class CsvParser extends ReportParser {
-
-        private static final long serialVersionUID = -8689695008930386640L;
+        private static final long serialVersionUID = 5363254965545196251L;
 
         @Override
-        public ReportDto parse(File file) {
-            return null;
+        public ReportDto parse(File file) throws IOException {
+            return new ObjectMapper(new XmlFactory()).readerFor(ReportDto.class).readValue(file);
         }
     }
 }
