+++ src/main/java/io/jenkins/plugins/reporter/nextgen/provider/Yaml.java
@@ -1,12 +1,12 @@
-package io.jenkins.plugins.reporter.steps.provider;
+package io.jenkins.plugins.reporter.nextgen.provider;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.xml.XmlFactory;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import edu.umd.cs.findbugs.annotations.CheckForNull;
 import hudson.Extension;
-import io.jenkins.plugins.reporter.steps.ReportDto;
-import io.jenkins.plugins.reporter.steps.Provider;
-import io.jenkins.plugins.reporter.steps.ReportParser;
+import io.jenkins.plugins.reporter.nextgen.ReportDto;
+import io.jenkins.plugins.reporter.nextgen.Provider;
+import io.jenkins.plugins.reporter.nextgen.ReportParser;
 import org.apache.commons.lang3.StringUtils;
 import org.jenkinsci.Symbol;
 import org.kohsuke.stapler.DataBoundConstructor;
@@ -15,18 +15,17 @@ import org.kohsuke.stapler.DataBoundSetter;
 import java.io.File;
 import java.io.IOException;
 
-public class Xml extends Provider {
+public class Yaml extends Provider {
     
     private static final long serialVersionUID = 9141170397250309265L;
 
-    private static final String ID = "xml";
-    
+    private static final String ID = "yaml";
     private String pattern = StringUtils.EMPTY;
-    
-    private String name = StringUtils.EMPTY;
 
+    private String name = StringUtils.EMPTY;
+    
     @DataBoundConstructor
-    public Xml() {
+    public Yaml() {
         super();
         // empty constructor required for stapler
     }
@@ -59,11 +58,11 @@ public class Xml extends Provider {
 
     @Override
     public ReportParser createParser() {
-        return new XmlParser();
+        return new YamlParser();
     }
 
     /** Descriptor for this provider. */
-    @Symbol("xml")
+    @Symbol({"yaml", "yml"})
     @Extension
     public static class Descriptor extends ProviderDescriptor {
         /** Creates the descriptor instance. */
@@ -71,14 +70,14 @@ public class Xml extends Provider {
             super(ID);
         }
     }
-    
-    public static class XmlParser extends ReportParser {
 
-        private static final long serialVersionUID = 5363254965545196251L;
+    public static class YamlParser extends ReportParser {
+
+        private static final long serialVersionUID = 8953162360286690397L;
 
         @Override
         public ReportDto parse(File file) throws IOException {
-            return new ObjectMapper(new XmlFactory()).readerFor(ReportDto.class).readValue(file);
+            return new ObjectMapper(new YAMLFactory()).readerFor(ReportDto.class).readValue(file);
         }
     }
 }
