+++ src/main/resources/io/jenkins/plugins/reporter/nextgen/JobAction/floatingBox.jelly
@@ -1,0 +1,6 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:c="/charts">
+    
+  JobAction View
+  
+</j:jelly>
