+++ src/main/resources/io/jenkins/plugins/reporter/nextgen/ReportDetails/index.jelly
@@ -1,0 +1,4 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:l="/lib/layout">
+    ReportDetails View
+</j:jelly>
